+++ x/perp/v2/keeper/grpc_query_test.go
@@ -453,3 +453,36 @@ func TestQueryPositionStore(t *testing.T) {
 
 	NewTestSuite(t).WithTestCases(tc...).Run()
 }
+
+func TestQueryCollateral(t *testing.T) {
+	tc := TestCases{
+		TC("state starts as expected with the mock NUSD denomination").
+			Given().
+			When().
+			Then(
+				QueryCollateral(types.TestingCollateralDenomNUSD),
+			),
+
+		TC("expected value returned after collateral denom changes").
+			Given().
+			When(
+				SetCollateral(denoms.BTC),
+			).
+			Then(
+				QueryCollateral(denoms.BTC),
+			),
+
+		TC("sanity check: multiple changes").
+			Given().
+			When(
+				SetCollateral(denoms.BTC),
+				SetCollateral(denoms.ETH),
+				SetCollateral(denoms.SOL),
+			).
+			Then(
+				QueryCollateral(denoms.SOL),
+			),
+	}
+
+	NewTestSuite(t).WithTestCases(tc...).Run()
+}