+++ x/perp/v2/types/msgs.go
@@ -40,10 +40,6 @@ func (m MsgRemoveMargin) ValidateBasic() error {
 		return fmt.Errorf("margin must be positive, not: %v", m.Margin.Amount.String())
 	}
 
-	if m.Margin.Denom != m.Pair.QuoteDenom() {
-		return fmt.Errorf("invalid margin denom, expected %s, got %s", m.Pair.QuoteDenom(), m.Margin.Denom)
-	}
-
 	return nil
 }
 
@@ -76,10 +72,6 @@ func (m MsgAddMargin) ValidateBasic() error {
 
 	if !m.Margin.Amount.IsPositive() {
 		return fmt.Errorf("margin must be positive, not: %v", m.Margin.Amount.String())
-	}
-
-	if m.Margin.Denom != m.Pair.QuoteDenom() {
-		return fmt.Errorf("invalid margin denom, expected %s, got %s", m.Pair.QuoteDenom(), m.Margin.Denom)
 	}
 
 	return nil
@@ -278,6 +270,33 @@ func (m MsgPartialClose) GetSignBytes() []byte {
 }
 
 func (m MsgPartialClose) GetSigners() []sdk.AccAddress {
+	signer, err := sdk.AccAddressFromBech32(m.Sender)
+	if err != nil {
+		panic(err)
+	}
+	return []sdk.AccAddress{signer}
+}
+
+// MsgChangeCollateralDenom
+
+func (m MsgChangeCollateralDenom) Route() string { return "perp" }
+func (m MsgChangeCollateralDenom) Type() string  { return "partial_close_msg" }
+
+func (m MsgChangeCollateralDenom) ValidateBasic() error {
+	if _, err := sdk.AccAddressFromBech32(m.Sender); err != nil {
+		return sdkerrors.Wrapf(errors.ErrInvalidAddress, "invalid sender address (%s)", err)
+	}
+	if err := sdk.ValidateDenom(m.NewDenom); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (m MsgChangeCollateralDenom) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
+}
+
+func (m MsgChangeCollateralDenom) GetSigners() []sdk.AccAddress {
 	signer, err := sdk.AccAddressFromBech32(m.Sender)
 	if err != nil {
 		panic(err)